+++ tests/test_keras.py
@@ -73,7 +73,7 @@ def test_basic_keras(dummy_model, dummy_data, wandb_init_run):
                     callbacks=[WandbCallback()])
     wandb.run.summary.load()
     assert wandb.run.history.rows[0]["epoch"] == 0
-    assert wandb.run.summary["acc"] > 0
+    assert wandb.run.summary["accuracy"] > 0
     assert len(wandb.run.summary["graph"].nodes) == 3
 
 
@@ -95,8 +95,16 @@ def test_basic_keras_multi_fit(dummy_model, dummy_data, wandb_init_run):
     assert wandb.run.history.rows[0]["epoch"] == 0
     assert wandb.run.history.rows[-1]["epoch"] == 1
     assert wandb.run.history.rows[-1]["_step"] == 3
-    assert wandb.run.summary["acc"] > 0
+    assert wandb.run.summary["accuracy"] > 0
     assert len(wandb.run.summary["graph"].nodes) == 3
+
+def test_keras_log_batch(dummy_model, dummy_data, wandb_init_run):
+    dummy_model.fit(*dummy_data, epochs=2, batch_size=10,
+                    callbacks=[WandbCallback(log_batch_frequency=2)])
+    wandb.run.summary.load()
+    print("History", wandb.run.history.rows)
+    assert len(wandb.run.history.rows) == 12
+
 
 
 def test_keras_image_bad_data(dummy_model, dummy_data, wandb_init_run):